pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
dependencies = ["dash>=2.18.2", "pillow"]
license = { text = "MIT License" }
name = "image_utils"
readme = "README.md"
requires-python = ">=3.10"
version = "0.0.1"
[project.optional-dependencies]
dev = ["image_utils[test]", "image_utils[docs]", "tox", "pre-commit"]
docs = ["sphinx", "myst-parser", "sphinx_rtd_theme"]
test = ["pytest", "pytest-cov", "coverage", "black", "isort", "pylint", "mypy", "ruff"]

[tool.setuptools.packages.find]
include = [
  "image_utils*",
] # package names should match these glob patterns (["*"] by default)
namespaces = false # to disable scanning PEP 420 namespaces (true by default)
where = ["src"] # list of folders that contain the packages (["."] by default)

[tool.isort]
line_length = 120
profile = "black"
skip = [".gitignore", ".dockerignore"]
skip_glob = ["docs/*", ".venv/*", "venv/*", ".tox/*"]

[tool.black]
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.pytest_cache
    | \.mypy_cache
    | \.venv
    | \.tox
    | venv
    | dist
  )/
)
'''
line_length = 120

[tool.pylint.messages_control]
disable = """
    missing-function-docstring,
    missing-module-docstring
    """
max-args = 10
max-line-length = 120
# generated-members = """
#     torch.*
#     numpy.*
#     """

[tool.pytest.ini_options]
addopts = "-v --cov=image_utils --cov-branch --cov-report html:reports/html --cov-report xml:reports/coverage.xml --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise AssertionError",
  "def __repr__",
  "def __str__",
  "if __name__ == .__main__.:",
  "raise NotImplementedError",
]

[tool.coverage.run]
branch = true
data_file = "reports/.coverage"

[tool.mypy]
strict = true
warn_return_any = false

[tool.ruff]
line-length = 120
